Splitting out saving utilities and globals into a dedicated utilities file.
